Translations and move TermsListField into FormTerms folder, refs ERM-683
