+++ lib/FormTerms/FormTerms.js
@@ -5,7 +5,7 @@ import { get } from 'lodash';
 
 import { Field } from 'react-final-form';
 
-import TermsListField from '../TermsListField';
+import TermsListField from './TermsListField';
 
 class FormTerms extends React.Component {
   static propTypes = {
@@ -31,7 +31,7 @@ class FormTerms extends React.Component {
           let options = get(term.category, ['values']);
           if (options) {
             options = [{
-              label: props.intl.formatMessage({ id: 'ui-licenses.notSet' }),
+              label: props.intl.formatMessage({ id: 'stripes-erm-components.formTerms.notSet' }),
               value: '',
             },
             ...options];
@@ -58,6 +58,7 @@ class FormTerms extends React.Component {
       <Field
         name="customProperties"
         validate={(value) => this.refToTermsListField.current && this.refToTermsListField.current.isInvalid(value)}
+        {...this.props}
         render={props => {
           return (
             <TermsListField
