+++ lib/FormTerms/TermsListField.js
@@ -27,6 +27,8 @@ export default class TermsListField extends React.Component {
       onChange: PropTypes.func,
     }),
     meta: PropTypes.object,
+    optionalSectionLabel: PropTypes.string.isRequired,
+    primarySectionLabel: PropTypes.string.isRequired,
     availableTerms: PropTypes.arrayOf(PropTypes.shape({
       description: PropTypes.string,
       label: PropTypes.string.isRequired,
@@ -96,7 +98,7 @@ export default class TermsListField extends React.Component {
         data-test-term-name
         dataOptions={[term, ...unsetTerms]} // The selected term, and the available unset terms
         id={`edit-term-${i}-name`}
-        label={<FormattedMessage id="ui-licenses.prop.termName" />}
+        label={<FormattedMessage id="stripes-erm-components.prop.termName" />}
         onChange={e => {
           const newValue = e.target.value;
 
@@ -141,14 +143,14 @@ export default class TermsListField extends React.Component {
       }
 
       if ((note && !value) || (publicNote && !value)) {
-        errors[term.value] = <FormattedMessage id="ui-licenses.errors.termNoteWithoutValue" />;
+        errors[term.value] = <FormattedMessage id="stripes-erm-components.errors.termNoteWithoutValue" />;
       }
 
       if (term.type === TERM_TYPE_NUMBER) {
         const min = Number.MIN_SAFE_INTEGER;
         const max = Number.MAX_SAFE_INTEGER;
         if (value < min || value > max) {
-          errors[term.value] = <FormattedMessage id="ui-licenses.errors.termValueNotInRange" values={{ min, max }} />;
+          errors[term.value] = <FormattedMessage id="stripes-erm-components.errors.termValueNotInRange" values={{ min, max }} />;
         }
       }
     });
@@ -198,7 +200,7 @@ export default class TermsListField extends React.Component {
       <FieldComponent
         data-test-term-value
         id={`edit-term-${i}-value`}
-        label={<FormattedMessage id="ui-licenses.prop.termValue" />}
+        label={<FormattedMessage id="stripes-erm-components.prop.termValue" />}
         onChange={handleChange}
         value={controlledFieldValue}
         error={!isEmpty(errors) ? errors[term.value] : undefined}
@@ -215,8 +217,8 @@ export default class TermsListField extends React.Component {
 
     const dataOptions = (intl) => {
       return [
-        { value: true, label: intl.formatMessage({ id: 'ui-licenses.term.internalTrue' }) },
-        { value: false, label: intl.formatMessage({ id: 'ui-licenses.term.internalFalse' }) }
+        { value: true, label: intl.formatMessage({ id: 'stripes-erm-components.term.internalTrue' }) },
+        { value: false, label: intl.formatMessage({ id: 'stripes-erm-components.term.internalFalse' }) }
       ];
     };
 
@@ -238,7 +240,7 @@ export default class TermsListField extends React.Component {
             data-test-term-visibility
             id={`edit-term-${i}-visibility`}
             dataOptions={dataOptions(intl)}
-            label={<FormattedMessage id="ui-licenses.prop.termVisibility" />}
+            label={<FormattedMessage id="stripes-erm-components.prop.termVisibility" />}
             onChange={handleChange}
             value={internal === undefined ? term.defaultInternal : internal}
           />
@@ -288,7 +290,7 @@ export default class TermsListField extends React.Component {
       <TextArea
         data-test-term-note
         id={`edit-term-${i}-internal-note`}
-        label={<FormattedMessage id="ui-licenses.term.internalNote" />}
+        label={<FormattedMessage id="stripes-erm-components.term.internalNote" />}
         onChange={handleChange}
         value={note}
       />
@@ -315,7 +317,7 @@ export default class TermsListField extends React.Component {
         data-test-term-public-note
         id={`edit-term-${i}-public-note`}
         onChange={handleChange}
-        label={<FormattedMessage id="ui-licenses.term.publicNote" />}
+        label={<FormattedMessage id="stripes-erm-components.term.publicNote" />}
         value={publicNote}
       />
     );
@@ -334,20 +336,21 @@ export default class TermsListField extends React.Component {
           });
         }}
       >
-        <FormattedMessage id="ui-licenses.terms.add" />
+        <FormattedMessage id="stripes-erm-components.terms.add" />
       </Button>
     );
   }
 
   renderTermsList = () => {
+    const { optionalSectionLabel, primarySectionLabel } = this.props;
     return (
       <div>
         <KeyValue
-          label={<FormattedMessage id="ui-licenses.terms.primaryTerms" />}
+          label={primarySectionLabel}
           value={this.renderTerms('primary')}
         />
         <KeyValue
-          label={<FormattedMessage id="ui-licenses.terms.optionalTerms" />}
+          label={optionalSectionLabel}
           value={this.renderTerms('optional')}
         />
       </div>
@@ -367,7 +370,7 @@ export default class TermsListField extends React.Component {
       } : null;
 
       const header = termType === 'optional' ?
-        <FormattedMessage id="ui-licenses.term.title" values={{ number: optionalTermCounter + 1 }} /> :
+        <FormattedMessage id="stripes-erm-components.term.title" values={{ number: optionalTermCounter + 1 }} /> :
         term.label;
 
       optionalTermCounter += 1;
@@ -376,7 +379,7 @@ export default class TermsListField extends React.Component {
         <EditCard
           data-test-term={termType}
           deleteBtnProps={deleteBtnProps}
-          deleteButtonTooltipText={<FormattedMessage id="ui-licenses.term.removeTerm" />}
+          deleteButtonTooltipText={<FormattedMessage id="stripes-erm-components.term.removeTerm" />}
           header={header}
           key={term.value}
           onDelete={termType === 'optional' ? () => this.handleDeleteTerm(term, i) : null}
